UIT-175: Fill in correct values for predefined filters
