+++ src/shop/_components/rewardsSection/RewardsSection.tsx
@@ -8,8 +8,8 @@ import { REWARD_TILE_WIDTH } from '../../../_components/reward/style';
 import { useTracking } from '../../../_context';
 import { TMainNavigationProp } from '../../../_routing';
 import { getRewardTrackingData } from '../../../_utils';
+import { TFilterRewardCategory, TFilterRewardSection, useRewardFilters } from '../../_hooks/useRewardFilters';
 import { useGetRewards } from '../../_queries/useGetRewards';
-import { getFiltersForCategory, getFiltersForSection, TFilterRewardCategory, TFilterRewardSection } from '../../_utils/reward';
 import { RewardsSectionLoader } from './RewardSection.loading';
 import * as Styled from './style';
 
@@ -33,6 +33,7 @@ export const RewardsSection = ({
   organizerId,
   ...props
 }: TRewardSectionProps) => {
+  const { getFiltersForCategory, getFiltersForSection } = useRewardFilters();
   const { data, isLoading } = useGetRewards({
     filters: { ...getFiltersForSection(section), ...getFiltersForCategory(category) },
     itemsPerPage: horizontal ? 20 : 3,
